requirements.txt
langchain
langchain-google-genai
beautifulsoup4
requests
pydantic
python-dotenv